[FIN-300] refactor: 아티클 미리보기 공통 타입 선언
