+++ components/SmartDrag/DragRelatedArticles/DragRelatedArticlesView.tsx
@@ -1,20 +1,12 @@
 import React, { useState, useEffect, useRef } from 'react';
 
 import IntroContainer from '@/components/SmartDrag/Intro/IntroView';
-import DragRelatedArticleBox from '@/components/SmartDrag/DragRelatedArticles/DragRelatedArticleBox';
+import DragRelatedArticleBox from '@/components/articles/ArticlePreview/DragRelatedArticleBox';
+import ArticlePreviewType from '@/types/Article';
 
 interface PropsType {
   selectedText: string;
-  articleList: {
-    id: number;
-    title: string;
-    body: string;
-    totalLike: number;
-    reply: number;
-    authorNickname: string;
-    date: string;
-    thumbnail: string;
-  }[];
+  articleList: ArticlePreviewType[];
   loadMoreItems: () => Promise<void>;
   page: number;
   setPage: React.Dispatch<React.SetStateAction<number>>;
